+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fix
 - Remove offline mode
+- Arc2D: mpl_plot
 
 ### Refactor
 - Implements InteractiveObject for handling all mouse objects in one class